Confbridge CLI new record file name check.
This fix checks to make sure that if a confbridge record start command is issued
from the CLI it will always use the file name given on the CLI even if it
changes between start/stop records for a conference.  Previously it had been
reusing the same file between start/stops even if a new filename was given.

(issue AST-1088)
Reported by: John Bigelow


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/11@382385 f38db490-d61c-443f-a65b-d21fe96a405b