Fixed spurious rebuilds of func_version.
func_version.so was being rebuilt every time, because build.h was
changing every build, because of the cleantest dependency that was
added in r384410 to fix parallel make bugs.

Now build.h will only be created if it does not exist, which was the
original behavior of the Makefile.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@384544 f38db490-d61c-443f-a65b-d21fe96a405b