Change several bridge functions to return error status.
The bridge frame queue functions need to return an error status if the
frame failed to be queued because of an error condition.  The main calls
that needed to return the status are:
ast_bridge_channel_queue_action_data() and
ast_bridge_channel_write_action_data().  The other return changes are
ripple effects.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@392435 f38db490-d61c-443f-a65b-d21fe96a405b