Change ARI user config to use a type field
When I initially wrote the configuration support for ARI users, I
determined the section type by a category prefix (i.e., [user-admin]).

This is neither idiomatic Asterisk configuration, nor is it really
that user friendly. This patch replaces the category prefix with a
type field in the section, which is much cleaner.

Review: https://reviewboard.asterisk.org/r/2664/


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@394076 f38db490-d61c-443f-a65b-d21fe96a405b