Add rtpengine configuration parameter
The rtpengine configuration parameter was documented in the XML documentation,
but it was not actually registered with the sorcery object. This adds the
parameter with a default of "asterisk", such that res_rtp_asterisk is chosen as
the default RTP implementation.

(closes issue ASTERISK-22380)
Reported by: Rusty Newton
Tested by: Rusty Newton


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@397621 f38db490-d61c-443f-a65b-d21fe96a405b