Fix bucket unit tests
After the review for buckets was completed (r2715), the handling of names in
the bucket core was deferred to the wizards. As such, the bucket unit tests
cannot expect that passing a URI with a scheme specified but no actual resource
name will automatically fail. The tests have been updated to not make this
check.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@397630 f38db490-d61c-443f-a65b-d21fe96a405b