Sanitize XML output for PIDF bodies.
PJSIP's PIDF API does not replace angle brackets with
their appropriate counterparts for XML. So we have to
do it ourself. In this particular case, the problem had
to do with attempting to place an unsanitized SIP URI
into an XML node. Now we don't get a 488 from recipients
of our PIDF NOTIFYs.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@397968 f38db490-d61c-443f-a65b-d21fe96a405b