Push CLI qualify into the threadpool.
If you run Asterisk in the background and then connect to
it through a separate console, the thread that runs CLI commands
is not registered with PJLIB. Thus PJLIB does not like it when
you attempt to send OPTIONS requests from that thread. So now
we push the task into the threadpool, which we know to be registered
with PJLIB.

Thanks to Antti Yrjola for reporting this.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@400680 f38db490-d61c-443f-a65b-d21fe96a405b