channels/chan_oss: Fix compilation problem on SmartOS/Illumos/SunOS
THis patch fixes an issue in chan_oss when building on certain platforms. It
ensures that soundcard.h is found.

Review: https://reviewboard.asterisk.org/r/3426

Note that this patch is a part of the patch on ASTERISK-23576; the Makefile
portion only applies to Asterisk 11+.

(issue ASTERISK-23576)
Reported by: Sebastian Wiedenroth
patches:
  fix-sunos.diff uploaded by Sebastian Wiedenroth (License 6597)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@412480 f38db490-d61c-443f-a65b-d21fe96a405b